Add a simple cache and 'by dir' suppport
The cache is just a pickle file of all the
data returned by scan_issues(). To use an
existing cache, run with '--cached' option.

The '--cached' option will load the data from
the cache instead of running scan_issues().
This is useful for testing template changes.

The data format is a pickle of the list returned
by scan_issues(). That list also now includes a
'dirs' item mapping directory names to a set of
issues referencing them.

scan_issues() now also returns a 'prs' object, but just
noticed that is a bit busted atm.